Falta Agregar lblProveedores

+++ Sistema de Gestion de Inventario Nuvéa/Modelos/Entidades/Producto.cs
@@ -17,7 +17,7 @@ namespace Modelos.Entidades
         private bool estado;
         private int cantidadStock;
         private long codigoBarras;
-        private double precioProduc;
+        private float precioProduc;
         private int idCategoria;
         private int idProveedor;
 
@@ -27,7 +27,7 @@ namespace Modelos.Entidades
         public bool Estado { get => estado; set => estado = value; }
         public int CantidadStock { get => cantidadStock; set => cantidadStock = value; }
         public long CodigoBarras { get => codigoBarras; set => codigoBarras = value; }
-        public double PrecioProduc { get => precioProduc; set => precioProduc = value; }
+        public float PrecioProduc { get => precioProduc; set => precioProduc = value; }
         public int IdCategoria { get => idCategoria; set => idCategoria = value; }
         public int IdProveedor { get => idProveedor; set => idProveedor = value; }
